refactor: move entropy main package to cmd/entropy
