--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/odpf/entropy/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
